requirements.txt
flask_wtf
gensim
nltk
bs4
spotipy
lxml
